requirements.txt
pandas
scipy
PyGithub
textstat
statsmodels
nltk